% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee2008}
\alias{lee2008}
\title{Lee (2008) Dataset}
\format{
A data frame with 6558 observations on the following 3 variables.
\describe{
  \item{y1}{a numeric vector (dependent variable). Democrat vote share election at t + 1.}
  \item{x1}{a numeric vector (running variable). The cutoff point is where x1 = 0. Democratic vote share at t.}
  \item{x2}{a numeric vector (additional covariate). Democratic vote share at t - 1}
}
}
\source{
Lee, D. S. 2008. Randomized experiments from non-random selection in U.S. House elections. Journal of Econometrics 142 (2): 675-697.
}
\usage{
data(lee2008)
}
\description{
A dataset used in Takahashi (2023).
}
\references{
Takahashi, M. 2023. Multiple imputation regression discontinuity designs: Alternative to regression discontinuity designs to estimate the local average treatment effect at the cutoff. Communications in Statistics - Simulation and Computation 52 (9): 4293-4312.
}
\keyword{datasets}
