% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LudwigMiller2007Modified}
\alias{LudwigMiller2007Modified}
\title{Ludwig and Miller (2007) Modified Dataset}
\format{
A data frame with 1037 observations on the following 11 variables.
\describe{
  \item{y1}{a numeric vector (dependent variable)}
  \item{x1}{a numeric vector (running variable). The cutoff point is where x1 = log(59.1984).}
  \item{z1}{a numeric vector (additional covariate)}
  \item{z2}{a numeric vector(additional covariate)}
  \item{z3}{a numeric vector(additional covariate)}
  \item{z4}{a numeric vector(additional covariate)}
  \item{z5}{a numeric vector(additional covariate)}
  \item{z6}{a numeric vector(additional covariate)}
  \item{z7}{a numeric vector(additional covariate)}
  \item{z8}{a numeric vector(additional covariate)}
  \item{z9}{a numeric vector(additional covariate)}
}
}
\source{
Ludwig, J., and D. L. Miller. 2007. Does Head Start improve children's life chances? Evidence from a regression discontinuity design. Quarterly Journal of Economics 122 (1): 159-208.
}
\usage{
data(LudwigMiller2007Modified)
}
\description{
A modified dataset used in Takahashi (2023).
}
\references{
Takahashi, M. 2023. Multiple imputation regression discontinuity designs: Alternative to regression discontinuity designs to estimate the local average treatment effect at the cutoff. Communications in Statistics - Simulation and Computation 52 (9): 4293-4312.
}
\keyword{datasets}
